Avoid calling addFunction after initialisation
In my use case 20% of runtime was spent calling emscriptens addFunction
which under the hood creates a new WebAsembly.Function each time.

Calling pushValue with a function will now store a reference to the
function in JS and creates a single wrapper function at instantiation
that will use the reference to the JS function from its upvalues.

The __gc metamethod has to be run on the userdata rather than the
function type because __gc only works for userdata types making
it a tad counter-intuitive that the pushed lua function doesn't
have the js_func metatable.

This should also decrease the number of objects for JS's GC to cleanup.